[CMAKE] Generate the windows version info file
Summary:
The version info resource files (`.rc`) are currently hardcoded and part
of the source tree but share most of their content.
This diffs makes the version info file a template and generates it as
needed. This will be later used to set populate version info for the dll
files.

Test Plan:
Build for windows and check the detailed properties for the binaries.
{F4434640}

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7569