[CMAKE] Drop dependency search for the native builds
Summary:
This is an alternative to D7552, intended to solve the need for
dependencies when building against the static depends.

This diff simply makes the dependency search conditional so it only
applies to the non native builds. It is limited to libevent and boost as
other dependencies already have an option to turn them off which
achieves the same.

Depends on D7816.

Test Plan:
On Ubuntu Xenial:
  cmake -GNinja .. -DCMAKE_TOOLCHAIN_FILE=../cmake/platforms/Linux64.cmake
  ninja
Check the build succeeds.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7720