[CMAKE] Lower boost test suites verbosity
Summary:
The boost tests are run individually in parallel by excluding all the
other tests. When a test fails, this causes Boost to print a list of the
skipped tests, and make locating the error difficult. This diff reduces
the verbosity level to prevent the skipped tests from being logged.

Test Plan:
Cause a test to fail, then run `ninja check` and ensure the skipped
tests log is gone.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D8785