[CMAKE] Prevent a false positive warning due to CMP0116
Summary:
The new policy doesn't know how to handle DEPFILE in custom command and warns if the call is located in a subdirectory, independenty of whether the code is compliant with the new policy or not. Our code is compliant, so just set this policy to NEW to remove the warning.
See https://gitlab.kitware.com/cmake/cmake/-/issues/21959#note_925225

Test Plan:
With CMake >= 3.20
  ninja
Check the warning is gone.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D9859