[secp256k1] build: Windows DLL additions
Summary:
Pass win32-dll to LT_INIT:

This is the recommended way to support building PE DLLs with modern
mingw toolchains and libtool.

> This option should be used if the package has been ported to build clean
> dlls on win32 platforms.
> If this macro is not used, libtool will assume that the package libraries
> are not dll clean and will build only static libraries on win32 hosts.

Add -no-undefined to libtool LDFLAGS:

Instruct libtool to not allow undefined symbols when linking a shared
library.

See:
https://autotools.io/libtool/windows.html
https://www.gnu.org/software/libtool/manual/libtool.html#LT_005fINIT
https://www.gnu.org/software/gnulib/manual/html_node/Libtool-and-Windows.html

This is a backport of [[https://github.com/bitcoin-core/secp256k1/pull/1022 | secp256k1#1022]]

The cmake change is based on secp256k1#1559 and is stronger than the autotools change, we pass the "-Wl,--no-undefined" whenever possible. The two cases in which it is not possible are:
- if external default callbacks are enabled, because the entire point of this option is that the user provides the callbacks at link time
- when using sanitizers, which involves a static sanitizer runtime (see discussion in  secp256k1#1559)

Test Plan: run the windows build on Cirrus-CI

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19469