[secp256k1] configure: Remove redundant pkg-config code
Summary:
This removes code that detects the pkg-config tool. We used this
back in the days when we had dependencies. ;) It can always be brought
back if we'll need it in the future.

Note that we still deliver a .pc file for this library, and there is
code in Makefile.am to install it. But this does not require the
pkg-config tool; only consumers of the .pc file will need it. This can
be verified by running `make install` (maybe after `mkdir /tmp/pre` and
`./configure --prefix=/tmp/pre` and checking that the .pc file is
installed correctly.

This is a backport of [[https://github.com/bitcoin-core/secp256k1/pull/1090 | secp256k1#1090]]

Test Plan:
```
./autogen.sh
mkdir build
cd build
../configure --prefix=/tmp/pre
make
make install
ls /tmp/pre/lib/pkgconfig/
```

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19544