[cmake] Fix a potential version mismatch in BerkeleyDB
Summary: If several version versions are installed, it is possible for the cmake find package to pick a header that doesn't match the library version. This diff fixes the issue by searching first for the most specific pattern as the library file name.

Test Plan:
On Archlinux with both db5.3 and db6.2 installed:
  ninja all check

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D12946