[secp256k1] tests: simplify `random_fe_non_zero` (remove loop limit and unneeded normalize)
Summary:
`random_fe_non_zero` contains a loop iteration limit that ensures that
we abort if `random_fe` ever yielded zero more than ten times in a row.
This construct was first introduced in PR #19 for
random non-square field elements and was later refactored into the
non-zero helper in PR #25. The copy-over to the
exhaustive tests happened recently in PR #1118.

This case seems to be practically irrelevant and I'd argue for keeping
things simple and removing it; if there's really a worry that the test's
random generator is heavily biased towards certain values or value
ranges then there should consequently be checks at other places too
(e.g. directly in `random_fe` for 256-bit values that repeatedly
overflow, i.e. >= p).

Also, the _fe_normalize call is not needed and can be removed, as the
result of `random_fe` is already normalized.

This is a partial backport of [[https://github.com/bitcoin-core/secp256k1/pull/1395 | secp256k1#1395]]
https://github.com/bitcoin-core/secp256k1/pull/1395/changes/dc5514144fb9d412aa3845432b053ee06a27da37

Test Plan: `ninja check-secp256k1`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D20154