[guix] Properly set the toolchain environment based on the CROSS_* environment variables
Summary:
These variables are used when building the static dependencies and need
to be used as well when cross compiling using cmake frome the GUIX
environment.

This ensures the proper toolchain is used and fixes the symbol check
failure in the GUIX build.

Test Plan: Run the GUIX build, check the symbol check completes with no error.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D15441