[nsis] Fix build error due to images not found
Summary: The NSIS installer installs the docs, i.e. the .md and .png files from the doc/ directory. Since the bitcoin_logo_doxygen.png file is excluded, there is no such file left since D16127. We could remove the lookup for images but it's more robust to just add a /nonfatal switch to avoid failing NSIS if no image is in the directory.

Test Plan:
  ./contrib/teamcity/build-configurations.py build-win64

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D16137