build: LLD based macOS toolchain
Summary:
```
This switches us to using a LLD based toolchain for macOS builds.
Benefits

 - Less complicated macOS toolchain.
 - No longer beholden to Apple releasing it's source for cctools, ld64 & libtapi.
 - No more reliance on third parties to modify those sources for us. i.e apple-libtapi, cctools-port (cctools + ld64).
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/21778 | core#21778]].

Note this needs some adaptation or our cmake build, as well as some update in build.sh to work around a cmake bug with llvm-strip.
The Cross build for OSX will be broken on CI (because it lacks the proper clang version). I paused the build for now and will unbreak in a follow-up.

Test Plan: Run the guix osx build.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17261